+++ Source/Bake/Cooking/Composers/DotNetComposer.cs
@@ -252,6 +252,7 @@ namespace Bake.Cooking.Composers
                     Platform.Any,
                     path,
                     ingredients.Version,
+                    properties,
                     new DirectoryArtifact(
                         Path.Combine(visualStudioProject.Directory, path)));
 
@@ -293,6 +294,7 @@ namespace Bake.Cooking.Composers
                     targetPlatform,
                     path,
                     ingredients.Version,
+                    properties,
                     new ExecutableArtifact(
                         visualStudioProject.CsProj.ToolCommandName,
                         Path.Combine(
@@ -367,8 +369,15 @@ namespace Bake.Cooking.Composers
             }
 
             var legacyVersion = ingredients.Version.LegacyVersion.ToString();
+
+            // CRITICAL for NuGet packages!
+            // Setting the AssemblyVersion to [MAJOR].0.0.0 ensures that the assembly version remains stable
+            // across minor and patch updates, which is important for binding redirects and compatibility.
+            properties["AssemblyVersion"] = $"{ingredients.Version.Major}.0.0.0";
+
             properties["Version"] = legacyVersion;
-            properties["AssemblyVersion"] = legacyVersion;
+            properties["ApplicationVersion"] = legacyVersion;
+            properties["FileVersion"] = legacyVersion;
             properties["AssemblyFileVersion"] = legacyVersion;
             properties["Description"] = BuildDescription(visualStudioSolution, ingredients);